+++ tests/testthat/test-chop.R
@@ -368,6 +368,20 @@ test_that("chop_spikes", {
 })
 
 
+test_that("isolate_chop", {
+  x <- c(1:2, rep(3, 3), 4:6)
+
+  expect_silent(res <- isolate_chop(x, breaks = c(2, 5), n = 2))
+  expect_equivalent(
+    res,
+    factor(c("[1, 2)", "[2, 5)", rep("{3}", 3), "[2, 5)", "[5, 6]", "[5, 6]"))
+  )
+
+  expect_silent(res2 <- isolate_chop(x, breaks = c(2, 5), prop = 0.25))
+  expect_equivalent(res, res2)
+})
+
+
 test_that("fillet", {
   x <- -2:2
   expect_silent(sole <- fillet(x, -1:1))